pyproject.toml
[project]
name = "digital-memoirs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "jsonify>=0.5",
    "netifaces>=0.11.0",
    "pillow>=11.3.0",
    "qrcode>=8.2",
    "requests>=2.32.5",
    "segno>=1.6.6",
    "sockets>=1.0.0",
    "uuid>=1.30",
    "watchdog>=6.0.0",
]
